// types/view.ts
/*
 * @FilePath: /mog-theme-iucky/types/view.ts
 * @author: Wibus
 * @Date: 2022-10-28 17:25:51
 * @LastEditors: Wibus
 * @LastEditTime: 2022-10-28 17:46:53
 * Coding With IU
 */

export interface ViewportRecord {
  w: number
  h: number
  mobile: boolean
  pad: boolean
  hpad: boolean
  wider: boolean
  widest: boolean
}